Add todo comment with respect to diffusive scaling
Currently the code aborts for various schemes, if the scaling is set
to acoustic!
By removing this configuration option, those would thereby essentially
disabled.
The better thing to do is to check whether we have to deal with a
multi-level and abort for those schemes only in that case, as they are
currently lacking multi-level support.