Read in scaling parameter again
After some discussion we now concluded that we continue reading
in the scaling parameter from the configuration, as we may want
to reintroduce it at some later point in time.
As it only is relevant in the case of multi-level meshes, we
throw an error when a diffusive scaling is requested for
multi-level meshes though.